- Code cleanup - Changed current to latest - Filtered status of 'inactive' from display
