+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@im-open/im-github-deployments",
-  "version": "1.0.0",
+  "version": "1.0.1",
   "main": "dist/index.esm.js",
   "types": "dist/index.d.ts",
   "license": "Apache-2.0",
