+++ src/components/Deployments/index.ts
@@ -1,2 +1,1 @@
-// export { Deployments } from './Deployments';
-export { Deployments2 as Deployments } from './Deployments2';
+export { Deployments } from './Deployments';
