+++ src/components/IMGitHubDeploymentsDashboard/GitHubDeploymentsComponents.tsx
@@ -3,11 +3,11 @@ import { useEntity } from '@backstage/plugin-catalog-react';
 import useAsyncRetry from 'react-use/lib/useAsyncRetry';
 import { RestDeployment, EnvDeployment } from '../../api/types';
 import { githubDeploymentsApiRef } from '../../api';
-// import { TableDeployments as Deployments } from '../Deployments/TableDeployments';
 import { Deployments } from '../Deployments';
 import { ResponseErrorPanel } from '@backstage/core-components';
 import { useApi } from '@backstage/core-plugin-api';
 import { Grid } from '@material-ui/core';
+import { DateTime } from 'ts-luxon';
 
 export const GitHubDeploymentsComponents = (props: {
   projectSlug: string;
@@ -55,10 +55,15 @@ export const GitHubDeploymentsComponents = (props: {
       deploymentNodeIds: ghDeployments.map(d => d.node_id),
     });
 
+    //get max deployment dates
+    // TODO: Need to figure out where *'s are coming from
+    const cleanEnv = (env: string) => env.toLowerCase().replace('*', '');
+
     for (var status of apiDeploymentStatuses) {
-      // TODO: Need to figure out where *'s are coming from
-      const statusEnvionment = status.environment.replace('*', '');
-      const env = getKeyFromList(statusEnvionment, catalogEnvironments);
+      const env = getKeyFromList(
+        cleanEnv(status.environment),
+        catalogEnvironments,
+      );
       const deployment = ghDeployments.filter(
         d => d.node_id == status.deployment_node_id,
       )[0];
@@ -68,14 +73,51 @@ export const GitHubDeploymentsComponents = (props: {
         displayEnvironment: env,
         state: status.state,
         proprojectSlug: projectSlug,
-        current: status.state.toLowerCase() == 'success',
+        latest: undefined,
       } as EnvDeployment);
     }
 
-    apiDeployments.sort((a, b) => (a.id < b.id ? 1 : -1));
+    const mapped = apiDeployments.reduce((acc: any, d: EnvDeployment) => {
+      const env = cleanEnv(d.environment);
+      const key = `${env}::${d.payload.instance}`;
+      acc[key] = acc[key] || [];
+      acc[key].push({
+        ...d,
+        environment: env,
+        instance: d.payload.instance,
+        created_at: d.created_at,
+      });
+      return acc;
+    }, {});
 
+    const toSeconds = (d: DateTime) =>
+      parseInt(DateTime.fromISO(d.toString()).toFormat('X'));
+
+    const maxDeployments = Object.entries(mapped).map(([_, value]) => {
+      const max = (value as any[]).reduce((prev, curr) =>
+        toSeconds(prev.created_at) >= toSeconds(curr.created_at) ? prev : curr,
+      );
+      return {
+        environment: max.environment,
+        instance: max.instance,
+        lastCreatedAt: max.created_at,
+      };
+    });
+
+    apiDeployments.sort((a, b) => (a.id < b.id ? 1 : -1));
     return {
-      apiDeployments: apiDeployments,
+      apiDeployments: apiDeployments.map(d => {
+        return {
+          ...d,
+          latest:
+            maxDeployments.filter(
+              f =>
+                f.environment == cleanEnv(d.environment) &&
+                f.instance == d.payload.instance &&
+                f.lastCreatedAt == d.created_at,
+            ).length > 0,
+        } as EnvDeployment;
+      }),
       catalogEnvironments: catalogEnvironments,
     };
   });
